Add context->scratchpad, for transient data
This came up in Topicbox, where we needed to keep track of a bunch of
updates during a single transaction, then do something at the end of it.

Each top-level transaction gets a new scratchpad, which is thrown away
when that transaction finishes.  This could maybe be an object, and have
better namespacing, but this works for now.