JMAP proc: be opaque in logs for unknown methods
Right now, if you try to call a method that doesn't exist, we'll happily
pass that through to record_call_info. This has led to some weird things
in our stats processor, where we log status for every API call by type.

I'm not crazy about the way this is done, but it seems right to do it in
Ix, and it's nice to have _some_ way of logging that an unknown method
was called. I suppose the other thing to do would just be not to record
it if it was an unknown method, but that seems less good.